+++ app/src/test/kotlin/me/xizzhu/android/joshua/ui/DialogHelper.kt
@@ -16,20 +16,20 @@
 
 package me.xizzhu.android.joshua.ui
 
-import android.content.Context
+import android.app.Activity
 import android.content.DialogInterface
 import androidx.annotation.StringRes
 
 object DialogHelper {
-    fun showDialog(context: Context, @StringRes title: Int, items: Array<String>, selected: Int,
+    fun showDialog(activity: Activity, @StringRes title: Int, items: Array<String>, selected: Int,
                    onClicked: DialogInterface.OnClickListener) {
     }
 
-    fun showDialog(context: Context, cancelable: Boolean, @StringRes message: Int,
+    fun showDialog(activity: Activity, cancelable: Boolean, @StringRes message: Int,
                    onPositive: DialogInterface.OnClickListener) {
     }
 
-    fun showDialog(context: Context, cancelable: Boolean, @StringRes message: Int,
+    fun showDialog(activity: Activity, cancelable: Boolean, @StringRes message: Int,
                    onPositive: DialogInterface.OnClickListener, onNegative: DialogInterface.OnClickListener?) {
     }
 }